# Optimizer that removes redundant DELIM_GET and DELIM_JOIN operators