--- a/src/execution/operator/join/physical_hash_join.cpp
+++ b/src/execution/operator/join/physical_hash_join.cpp
@@ -10,9 +10,9 @@ namespace duckdb {
 
 PhysicalHashJoin::PhysicalHashJoin(LogicalOperator &op, unique_ptr<PhysicalOperator> left,
                                    unique_ptr<PhysicalOperator> right, vector<JoinCondition> cond, JoinType join_type,
-                                   vector<idx_t> left_projection_map, vector<idx_t> right_projection_map)
+                                   vector<idx_t> left_projection_map, vector<idx_t> right_projection_map, vector<LogicalType> delim_types)
     : PhysicalComparisonJoin(op, PhysicalOperatorType::HASH_JOIN, move(cond), join_type),
-      right_projection_map(right_projection_map) {
+      right_projection_map(right_projection_map), delim_types(move(delim_types)) {
 	children.push_back(move(left));
 	children.push_back(move(right));
 
@@ -29,7 +29,7 @@ PhysicalHashJoin::PhysicalHashJoin(LogicalOperator &op, unique_ptr<PhysicalOpera
 
 PhysicalHashJoin::PhysicalHashJoin(LogicalOperator &op, unique_ptr<PhysicalOperator> left,
                                    unique_ptr<PhysicalOperator> right, vector<JoinCondition> cond, JoinType join_type)
-    : PhysicalHashJoin(op, move(left), move(right), move(cond), join_type, {}, {}) {
+    : PhysicalHashJoin(op, move(left), move(right), move(cond), join_type, {}, {}, {}) {
 }
 
 //===--------------------------------------------------------------------===//