--- a/src/execution/operator/join/physical_hash_join.cpp
+++ b/src/execution/operator/join/physical_hash_join.cpp
@@ -10,7 +10,8 @@ namespace duckdb {
 
 PhysicalHashJoin::PhysicalHashJoin(LogicalOperator &op, unique_ptr<PhysicalOperator> left,
                                    unique_ptr<PhysicalOperator> right, vector<JoinCondition> cond, JoinType join_type,
-                                   vector<idx_t> left_projection_map, vector<idx_t> right_projection_map, vector<LogicalType> delim_types)
+                                   vector<idx_t> left_projection_map, vector<idx_t> right_projection_map,
+                                   vector<LogicalType> delim_types)
     : PhysicalComparisonJoin(op, PhysicalOperatorType::HASH_JOIN, move(cond), join_type),
       right_projection_map(right_projection_map), delim_types(move(delim_types)) {
 	children.push_back(move(left));