--- a/src/include/duckdb/optimizer/deliminator.hpp
+++ b/src/include/duckdb/optimizer/deliminator.hpp
@@ -22,9 +22,9 @@ public:
 
 private:
 	//! TODO
-    void FindCandidates(unique_ptr<LogicalOperator> *op_ptr, vector<unique_ptr<LogicalOperator> *> &candidates);
+	void FindCandidates(unique_ptr<LogicalOperator> *op_ptr, vector<unique_ptr<LogicalOperator> *> &candidates);
 	//! TODO
-    bool RemoveCandidate(unique_ptr<LogicalOperator> *op_ptr, column_binding_map_t<unique_ptr<Expression>> &mapping);
+	bool RemoveCandidate(unique_ptr<LogicalOperator> *op_ptr, column_binding_map_t<unique_ptr<Expression>> &mapping);
 	//! Replace references to a removed DelimGet, remove DelimJoins if all their DelimGets are gone
 	void UpdatePlan(LogicalOperator &op, column_binding_map_t<unique_ptr<Expression>> &cb_map);
 };