--- a/src/include/duckdb/optimizer/deliminator.hpp
+++ b/src/include/duckdb/optimizer/deliminator.hpp
@@ -24,9 +24,9 @@ private:
 	//! TODO
 	void FindCandidates(unique_ptr<LogicalOperator> *op_ptr, vector<unique_ptr<LogicalOperator> *> &candidates);
 	//! TODO
-	bool RemoveCandidate(unique_ptr<LogicalOperator> *op_ptr, expression_map_t<Expression *> &expr_map, unique_ptr<LogicalOperator> *temp_ptr);
+	bool RemoveCandidate(unique_ptr<LogicalOperator> *op_ptr, expression_map_t<Expression *> &expr_map, column_binding_map_t<bool> &projection_map, unique_ptr<LogicalOperator> *temp_ptr);
 	//! Replace references to a removed DelimGet, remove DelimJoins if all their DelimGets are gone
-	void UpdatePlan(LogicalOperator &op, expression_map_t<Expression *> &expr_map);
+	void UpdatePlan(LogicalOperator &op, expression_map_t<Expression *> &expr_map, column_binding_map_t<bool> &projection_map);
 };
 
 } // namespace duckdb