--- a/src/include/duckdb/optimizer/deliminator.hpp
+++ b/src/include/duckdb/optimizer/deliminator.hpp
@@ -17,18 +17,20 @@ class Deliminator {
 public:
 	Deliminator() {
 	}
-
+	//! Perform DelimJoin elimination
 	unique_ptr<LogicalOperator> Optimize(unique_ptr<LogicalOperator> op);
 
 private:
-	//! TODO
+	//! Find Joins with a DelimGet that can be removed
 	void FindCandidates(unique_ptr<LogicalOperator> *op_ptr, vector<unique_ptr<LogicalOperator> *> &candidates);
-	//! TODO
+	//! Try to remove a Join with a DelimGet, returns true if it was successful
 	bool RemoveCandidate(unique_ptr<LogicalOperator> *op_ptr, expression_map_t<Expression *> &expr_map,
 	                     column_binding_map_t<bool> &projection_map, unique_ptr<LogicalOperator> *temp_ptr);
 	//! Replace references to a removed DelimGet, remove DelimJoins if all their DelimGets are gone
 	void UpdatePlan(LogicalOperator &op, expression_map_t<Expression *> &expr_map,
 	                column_binding_map_t<bool> &projection_map);
+	//! Whether the operator has one or more children of type DELIM_GET
+    bool HasChildDelimGet(LogicalOperator &op);
 };
 
 } // namespace duckdb