--- a/src/include/duckdb/optimizer/deliminator.hpp
+++ b/src/include/duckdb/optimizer/deliminator.hpp
@@ -11,6 +11,7 @@
 namespace duckdb {
 
 class Optimizer;
+class DeliminatorPlanUpdater;
 
 //! The Deliminator optimizer traverses the logical operator tree and removes any redundant DelimGets/DelimJoins
 class Deliminator {
@@ -24,8 +25,7 @@ private:
 	//! Find Joins with a DelimGet that can be removed
 	void FindCandidates(unique_ptr<LogicalOperator> *op_ptr, vector<unique_ptr<LogicalOperator> *> &candidates);
 	//! Try to remove a Join with a DelimGet, returns true if it was successful
-	bool RemoveCandidate(unique_ptr<LogicalOperator> *op_ptr, expression_map_t<Expression *> &expr_map,
-	                     column_binding_map_t<bool> &projection_map, unique_ptr<LogicalOperator> *temp_ptr);
+	bool RemoveCandidate(unique_ptr<LogicalOperator> *op_ptr, DeliminatorPlanUpdater &updater);
 	//! Replace references to a removed DelimGet, remove DelimJoins if all their DelimGets are gone
 	void UpdatePlan(LogicalOperator &op, expression_map_t<Expression *> &expr_map,
 	                column_binding_map_t<bool> &projection_map);