--- a/src/include/duckdb/planner/operator/logical_comparison_join.hpp
+++ b/src/include/duckdb/planner/operator/logical_comparison_join.hpp
@@ -22,7 +22,7 @@ public:
 
 	//! The conditions of the join
 	vector<JoinCondition> conditions;
-	//! Used for duplicate-eliminated TODO
+	//! Used for duplicate-eliminated joins
 	vector<LogicalType> delim_types;
 
 public: