--- a/src/include/duckdb/planner/operator/logical_comparison_join.hpp
+++ b/src/include/duckdb/planner/operator/logical_comparison_join.hpp
@@ -18,7 +18,8 @@ namespace duckdb {
 //! LogicalComparisonJoin represents a join that involves comparisons between the LHS and RHS
 class LogicalComparisonJoin : public LogicalJoin {
 public:
-	LogicalComparisonJoin(JoinType type, LogicalOperatorType logical_type = LogicalOperatorType::LOGICAL_COMPARISON_JOIN);
+	LogicalComparisonJoin(JoinType type,
+	                      LogicalOperatorType logical_type = LogicalOperatorType::LOGICAL_COMPARISON_JOIN);
 
 	//! The conditions of the join
 	vector<JoinCondition> conditions;