… src/optimizer/CMakeLists.txt
… src/optimizer/optimizer.cpp
… src/planner/binder/query_node/plan_subquery.cpp
… src/planner/subquery/flatten_dependent_join.cpp
… test/optimizer/deliminator_optimizer.test_slow