@Mytherin 2021-01-18
Thanks, looks great now!